+++ src/shared/constants.ts
@@ -1,11 +1,27 @@
 export const headers = {
-  Referer: "https://trading.vietcap.com.vn/price-board?filter-group=WL&filter-value=DEFAULT&view-type=FLAT",
+  Accept: "application/json, text/plain, */*",
+  "Accept-Language": "en-US,en;q=0.9,vi-VN;q=0.8,vi;q=0.7",
+  "Content-Type": "application/json",
+  Connection: "keep-alive",
+  "Cache-Control": "no-cache",
+  Pragma: "no-cache",
+  DNT: "1",
+  Origin: "https://trading.vietcap.com.vn",
+  Referer: "https://trading.vietcap.com.vn/",
 };
 
 export const BASE_URL = "https://trading.vietcap.com.vn";
+export const VCI_TRADING_URL = BASE_URL + "/api";
+export const VCI_HANDSHAKE_URL = BASE_URL + "/priceboard";
 export const CHART_URL = BASE_URL + "/api/chart/OHLCChart/gap-chart";
 export const ALL_SYMBOLS_URL = "https://ai.vietcap.com.vn/api/get_all_tickers";
+
 export const GRAPHQL_URL = BASE_URL + "/data-mt/graphql";
+
+export const VCI_IQ_URL = "https://iq.vietcap.com.vn/api/iq-insight-service";
+export const VCI_COMPANY_URL = VCI_IQ_URL + "/v1/company";
+export const VCI_EVENTS_URL = VCI_IQ_URL + "/v1/events";
+export const VCI_NEWS_URL = VCI_IQ_URL + "/v1/news";
 
 export const INTERVAL_MAP = {
   "1m": "ONE_MINUTE",